pyproject.toml
[tool.poetry]
name = "tg-summarizer"
version = "0.1.0"
description = ""
authors = ["Konstantin Litvin <konstantinlitvin89@gmail.com>"]
readme = "README.md"
packages = [{include = "tg_summarizer"}]

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.28.5"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
